signOut, translations and typescript left
