+++ frontend/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "@babel/runtime": "7.5.4",
-    "@cthit/react-digit-components": "1.0.2",
+    "@cthit/react-digit-components": "^1.0.3-beta.12",
     "@material-ui/core": "4.9.11",
     "@material-ui/icons": "4.9.1",
     "axios": "0.19.2",
