+++ frontend/src/api/gamma-signout/post.gamma-signout.api.js
@@ -3,5 +3,4 @@ import { GAMMA_SIGNOUT_ENDPOINT } from "../utils/endpoints";
 
 export function signoutFromSongbook() {
     postRequest(GAMMA_SIGNOUT_ENDPOINT);
-    window.location.reload();
 }
