+++ frontend/src/api/utils/api.js
@@ -1,8 +1,6 @@
 import axios from "axios";
 import _ from "lodash";
 
-const path = "http://localhost:5000/";
-
 export function getRequest(endpoint) {
     return axios.get(removeLastSlash(endpoint));
 }
