+++ frontend/src/app/Elements/Header/index.js
@@ -1,0 +1,2 @@
+import Header from "./HeaderElement.element";
+export default Header;
