+++ frontend/src/common-ui/design/Common.styles.jsx
@@ -1,30 +1,28 @@
 import styled from "styled-components";
-import { DigitDesign, DigitLoading, DigitTextField } from "@cthit/react-digit-components";
+import { DigitDesign } from "@cthit/react-digit-components";
 
 export const ColumnContainer = styled.div`
-  display: flex;
-  width: 100vw;
-  flex-direction: column;
+    display: flex;
+    width: 100vw;
+    flex-direction: column;
 `;
 
-
 export const CenterContainer = styled.div`
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
 `;
 
 export const WideCenterContainer = styled(CenterContainer)`
     width: min(80vw, 600px);
-`
+`;
 
 export const TopRightButton = styled.div`
-  width: 100vw;
-`
+    width: 100vw;
+`;
 
 export const DangerCard = styled(DigitDesign.Card)`
     background-color: #e3242b;
     margin: 15px;
     max-width: 500px;
-`
+`;
