+++ frontend/src/common-ui/layout/CentralLoading.jsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { CenterContainer } from "./Common.styles";
+import { CenterContainer } from "../design/Common.styles";
 import { DigitLoading } from "@cthit/react-digit-components";
 
 export const CentralLoading = () => (
     <CenterContainer>
-        <DigitLoading/>
+        <DigitLoading />
     </CenterContainer>
-)
+);
