+++ frontend/src/use-cases/songs/view-songs/index.js
@@ -1,0 +1,3 @@
+import ViewSongs from "./View-Songs.Screen";
+
+export default ViewSongs;
