+++ frontend/src/use-cases/songs/view-songs/views/views.styles.jsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { DigitDesign, DigitTextField } from "@cthit/react-digit-components";
 
 export const FilterBody = styled.div`
     display: flex;
